优化
配合stm32_define,优化外设访问方式

+++ mcu/stm32f0x/ebox/Peripherals/ebox_gpio.h
@@ -41,7 +41,7 @@ public:
     E_PinID _id;
 
     E_PinBase(E_PinID pin);
-    void mode(E_PinMode mode,uint16_t af);
+    void mode(E_PinMode mode,uint8_t af);
 };
 
 /*+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++*/
